Identify some interesting patterns and their reverse
Also force a multiplier of 0.7 between speed and pattern speed
(temporary address)

I thought that the speed might come as negative and it might be possible
to set the pattern rotate in reverse when riding backwards, but EUC World
(and maybe the wheel itself) always shows positive wheel values.
So for now, it's not too useful, but I kept the code for now.